Prettify TaskError usage and add Task module for future changes
