+++ proxy/src/lib.rs
@@ -48,12 +48,7 @@ extern crate tower_in_flight_limit;
 extern crate trust_dns_resolver;
 
 use futures::*;
-use futures::future::{
-    ExecuteError,
-    ExecuteErrorKind,
-    Executor
-};
-
+use futures::future::Executor;
 use std::error::Error;
 use std::fmt;
 use std::io;
@@ -67,7 +62,6 @@ use tokio::{
     executor::{
         current_thread::{self, CurrentThread},
         thread_pool::{self, Sender},
-        SpawnError,
     },
     reactor,
 };
@@ -88,6 +82,7 @@ mod inbound;
 mod logging;
 mod map_err;
 mod outbound;
+mod task;
 pub mod telemetry;
 mod transparency;
 mod transport;
@@ -98,6 +93,7 @@ use bind::Bind;
 use connection::BoundPort;
 use inbound::Inbound;
 use map_err::MapErr;
+use task::TaskError;
 use transparency::{HttpBody, Server};
 pub use transport::{GetOriginalDst, SoOriginalDst};
 use outbound::Outbound;
@@ -127,21 +123,6 @@ pub struct Main<G> {
 
     // runtime: CurrentThread<tokio_timer::Timer<tokio_reactor::Reactor>>,
     // handle: reactor::Handle,
-}
-
-/// Like a `SpawnError` or `ExecuteError`, but with an implementation
-/// of `std::error::Error`.
-#[derive(Debug, Clone)]
-pub enum TaskError {
-    /// The executor has shut down and will no longer spawn new tasks.
-    Shutdown,
-    /// The executor had no capacity to run more futures.
-    NoCapacity,
-    /// An unknown error occurred.
-    ///
-    /// This indicates that `tokio` or `futures-rs` has
-    /// added additional error types that we are not aware of.
-    Unknown,
 }
 
 impl<G> Main<G>
@@ -579,62 +560,8 @@ where
 
             future::result(executor.execute(s)
                 .map(move |_| (server, executor))
-                .map_err(|e| io::Error::new(io::ErrorKind::Other, TaskError::from(e)))
+                .map_err(TaskError::into_io)
             )
         },
     )
-}
-
-//===== impl TaskError =====
-
-impl From<SpawnError> for TaskError {
-    fn from(spawn_error: SpawnError) -> Self {
-        if spawn_error.is_shutdown() {
-            TaskError::Shutdown
-        } else if spawn_error.is_at_capacity() {
-            TaskError::NoCapacity
-        } else {
-            warn!(
-                "TaskError::from: unknown SpawnError '{:?}'\n\
-                 This indicates a change in Tokio's API surface that should\n\
-                 be handled.",
-                 spawn_error,
-            );
-            TaskError::Unknown
-        }
-    }
-}
-
-impl<F> From<ExecuteError<F>> for TaskError {
-    fn from(exec_error: ExecuteError<F>) -> Self {
-        match exec_error.kind() {
-            ExecuteErrorKind::Shutdown => TaskError::Shutdown,
-            ExecuteErrorKind::NoCapacity => TaskError::NoCapacity,
-            _ => {
-                warn!(
-                    "TaskError::from: unknown ExecuteError '{:?}'\n\
-                    This indicates a change in the futures-rs API surface\n\
-                    that should be handled.",
-                    exec_error,
-                );
-                TaskError::Unknown
-            }
-        }
-    }
-}
-impl fmt::Display for  TaskError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.description().fmt(f)
-    }
-}
-
-impl Error for TaskError {
-    fn description(&self) -> &str {
-        match *self {
-            TaskError::Shutdown => "executor has shut down",
-            TaskError::NoCapacity => "executor has no more capacity",
-            TaskError::Unknown => "unknown error executing future",
-        }
-    }
-
 }
